fix: separate particle engine code
